Cleanup (seperate classes into logical files)
