+++ HydrateOrDiedrate/src/Config/Sync/ConfigSyncRequestPacket.cs
@@ -1,0 +1,10 @@
+﻿using ProtoBuf;
+
+namespace HydrateOrDiedrate.src.Config.Sync;
+
+[ProtoContract]
+public class ConfigSyncRequestPacket
+{
+
+}
+
